+++ Plugins/ChronicleEngine/Source/ChronicleEngineEditor/Public/Asset/ChronicleDialogueJsonLibrary.h
@@ -1,10 +1,51 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "GameplayTagContainer.h"
 #include "Kismet/BlueprintFunctionLibrary.h"
 #include "ChronicleDialogueJsonLibrary.generated.h"
 
 class UDialogueTree;
+class UDialogueDatabase;
+
+USTRUCT(BlueprintType)
+struct CHRONICLEENGINEEDITOR_API FChronicleDialogueLocalizationEntry
+{
+    GENERATED_BODY()
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FString Namespace;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FName Key;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FString Culture;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FGuid TreeGuid;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FGuid NodeGuid;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    int32 LineIndex = INDEX_NONE;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FName LineID;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FGameplayTag SpeakerTag;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FText SourceText;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FText TranslatedText;
+
+    UPROPERTY(BlueprintReadOnly, Category="Chronicle|Localization")
+    FString ContextComment;
+};
 
 UENUM(BlueprintType)
 enum class EChronicleDialogueValidationSeverity : uint8
@@ -58,6 +99,21 @@ public:
 
     UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|CSV")
     static bool ImportDialogueLinesFromCsvFile(UDialogueTree* Tree, const FString& FilePath, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Localization")
+    static bool EnsureStableLineIds(UDialogueTree* Tree, int32& OutUpdatedCount, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Localization")
+    static bool GatherDialogueTextsFromTree(UDialogueTree* Tree, const FString& Namespace, const FString& Culture, TArray<FChronicleDialogueLocalizationEntry>& OutEntries, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Localization")
+    static bool GatherDialogueTextsFromDatabase(UDialogueDatabase* Database, const FString& Culture, TArray<FChronicleDialogueLocalizationEntry>& OutEntries, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Localization")
+    static bool ExportLocalizationCsvFromTree(UDialogueTree* Tree, const FString& Namespace, const FString& Culture, FString& OutCsv, FString& OutError);
+
+    UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Localization")
+    static bool ImportLocalizationCsvToTree(UDialogueTree* Tree, const FString& Csv, FString& OutError);
 
     UFUNCTION(BlueprintCallable, Category="Chronicle|Editor|Validation")
     static bool ValidateDialogueTree(UDialogueTree* Tree, TArray<FChronicleDialogueValidationIssue>& OutIssues);